linter & cooldowns
oh yes

+++ commands/Utilities/say.js
@@ -12,19 +12,15 @@ module.exports = {
 
     run: async (bot, message, args) => {
 
-        if (message.deletable) {
-                message.delete();
-            }
-            if (args.length < 1) {
-                return (await message.reply("Bruh you didn't even give me a message, what am I supposed to do, read your mind?"))
-            }
-            if (args[0].toLowerCase() === 'embed') {
-                const sayEmbed = new Discord.MessageEmbed()
-                    .setColor('#228b22')
-                    .setDescription(args.slice(1).join(' '));
-                message.channel.send(sayEmbed);
-            } else {
-                message.channel.send(args.join(' '));
-            }
-}
-}
+        if (message.deletable) message.delete();
+        if (args.length < 1) return (await message.send('Bruh you didn\'t even give me a message, what am I supposed to do, read your mind?'));
+        if (args[0].toLowerCase() === 'embed') {
+            const sayEmbed = new Discord.MessageEmbed()
+                .setColor('#228b22')
+                .setDescription(args.slice(1).join(' '));
+            message.channel.send(sayEmbed);
+        } else {
+            message.channel.send(args.join(' '));
+        }
+    }
+};
